requirements.txt
aioboto3==13.3.0
aiobotocore==2.16.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aioredis==1.3.1
aioredlock==0.7.3
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
asyncache==0.3.1
asyncpg==0.30.0
attrs==24.3.0
autoflake==2.3.1
Automat==24.8.1
babel==2.16.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
boto3==1.35.81
botocore==1.35.81
Brotli==1.1.0
browsermob-proxy==0.8.0
browser-use==0.1.40
bs4==0.0.2
build==1.2.2.post1
bytecode==0.16.0
CacheControl==0.14.1
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
cheap_repr==0.5.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
cleo==2.1.0
click==8.1.8
clipboard==0.0.4
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
commentjson==0.9.0
constantly==23.10.4
crashtest==0.4.1
cryptography==44.0.0
cssselect==1.2.0
curlify==2.2.1
ddsketch==3.0.1
ddtrace==2.18.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
dulwich==0.21.7
durationpy==0.9
ecdsa==0.19.0
email_validator==2.2.0
envier==0.6.1
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
filetype==1.2.0
flake8==7.1.1
Flask==3.0.3
flatbuffers==24.3.25
fpdf==1.7.2
fqdn==1.5.1
freezegun==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
gotrue==2.11.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h2==4.1.0
hiredis==3.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.27.0
humanfriendly==10.0
hyperframe==6.0.1
hyperlink==21.0.0
identify==2.6.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==24.7.2
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.13.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaitaistruct==0.10
keyring==24.3.1
kubernetes==31.0.0
lark-parser==0.7.8
litellm==1.55.9
langchain==0.3.20
langchain-anthropic==0.3.3
langchain-core==0.3.41
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.6
lxml==5.3.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.1
mock==5.1.0
monotonic==1.6
more-itertools==10.5.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
mss==10.0.0
multidict==6.1.0
mypy==1.14.0
mypy-extensions==1.0.0
narwhals==1.19.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.1
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.58.1
opencv-python==4.10.0.84
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
pkginfo==1.12.0
platformdirs==4.3.6
playwright==1.49.1
plotly==5.24.1
pluggy==1.5.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
postgrest==0.18.0
posthog==3.7.4
pre_commit==4.0.1
primp==0.9.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydeck==0.9.1
pydevd-pycharm==243.22562.220
pynput==1.7.7
PyDispatcher==2.0.7
pyee==12.0.0
pyflakes==3.2.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyOpenSSL==24.3.0
pyotp==2.9.0
pyparsing==3.2.0
pypdf==5.1.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-playwright==0.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==3.2.1
python-multipart==0.0.20
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.7.0
RapidFuzz==3.11.0
realtime==2.0.6
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
scapy==2.6.1
Scrapy==2.12.0
SecretStorage==3.3.3
selenium==4.27.1
selenium-wire==5.1.0
Send2Trash==1.8.3
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
snoop==0.6.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.41.3
starlette-context==0.3.6
storage3==0.9.0
streamlit==1.41.1
StrEnum==0.4.15
stripe==11.4.1
structlog==24.4.0
supabase==2.10.0
supafunc==0.7.0
sympy==1.13.3
temporalio==1.8.0
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
tldextract==5.1.3
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
trove-classifiers==2024.10.21.16
Twisted==24.11.0
typer==0.15.1
types-protobuf==5.29.1.20241207
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
types-toml==0.10.8.20240310
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.0
w3lib==2.2.1
watchdog==6.0.0
watchfiles==1.0.3
waybackpy==3.0.6
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.0
wsproto==1.2.0
xmltodict==0.14.2
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
